ASoC: msm: qdsp6v2: support multi-ch and samplerate > 48K for NT wma
NT and compress offload wma mode are exercised, we need to implement
changes in framework so that both have the same restrictions.

Meanwhile, we will remove any limitation in driver to support
multichannel and sample rate more than 48KHz wma playback in NT path.

Add extra option to support default channel mapping for both
back-speaker and side-speaker flavor.

CRs-Fixed: 870824

+++ include/uapi/linux/msm_audio_wmapro.h
@@ -7,16 +7,16 @@
 	  (AUDIO_MAX_COMMON_IOCTL_NUM+1), struct msm_audio_wmapro_config)
 
 struct msm_audio_wmapro_config {
-	unsigned short 	armdatareqthr;
+	unsigned short  armdatareqthr;
 	uint8_t         validbitspersample;
 	uint8_t         numchannels;
 	unsigned short  formattag;
-	unsigned short  samplingrate;
-	unsigned short  avgbytespersecond;
+	uint32_t        samplingrate;
+	uint32_t        avgbytespersecond;
 	unsigned short  asfpacketlength;
-	unsigned short 	channelmask;
-	unsigned short 	encodeopt;
-	unsigned short	advancedencodeopt;
-	uint32_t	advancedencodeopt2;
+	uint32_t        channelmask;
+	unsigned short  encodeopt;
+	unsigned short  advancedencodeopt;
+	uint32_t        advancedencodeopt2;
 };
 #endif /* _UAPI_MSM_AUDIO_WMAPRO_H */
